Use Longs as Task IDs

+++ app/src/main/kotlin/de/robinwersich/todue/data/entities/Task.kt
@@ -7,7 +7,7 @@ import java.time.LocalDate
 
 @Entity(tableName = "todo")
 data class Task(
-  @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Int = 0,
+  @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Long = 0,
   @ColumnInfo(name = "text") val text: String,
   @ColumnInfo(name = "due_date") val dueDate: LocalDate,
   @ColumnInfo(name = "done_date") val doneDate: LocalDate? = null
