+++ app/src/main/kotlin/de/robinwersich/todue/ui/components/TaskEvent.kt
@@ -2,9 +2,9 @@ package de.robinwersich.todue.ui.components
 
 sealed class TaskEvent {
   object Add : TaskEvent()
-  data class Remove(val id: Int) : TaskEvent()
-  data class Expand(val id: Int) : TaskEvent()
+  data class Remove(val id: Long) : TaskEvent()
+  data class Expand(val id: Long) : TaskEvent()
   object Collapse : TaskEvent()
-  data class SetText(val id: Int, val text: String) : TaskEvent()
-  data class SetDone(val id: Int, val done: Boolean) : TaskEvent()
+  data class SetText(val id: Long, val text: String) : TaskEvent()
+  data class SetDone(val id: Long, val done: Boolean) : TaskEvent()
 }
